refactor: move init/cleanup methods to Landscape
